SONG Sequence Read returns null instead of empty strings Some fields for entities imported from SONG were indexed with facet values of empty strings, instead of missing data.
The culprit is a wrapper on Jackson JSON object that would get() a String value using the asText() method which opts to return empty string when the key/value is missing.

By modifying the get() method in SongSequencingRead we make it so that all usages in the SongProcessor will get null (gets indexed as empty value) instead of "" (empty string which gets indexed wrong).

This affects the values retrieved for:
 - experimental strategy
 - reference genome build
 - analysis software