Collab is now a wrapper for SONG, and SONG has become repository agnostic SONG is a repository API tool for DCC. It has been implemented in the Collab repository only so far but could be used in other places. The object here is to make SONG generic and make collaboratories that use the SONG API be an implementing wrapper over SONG.
To do this, we made the SongImporter abstract and removed the context only constructor that is used for all other repositories. This constructor was implemented in the CollabImporter which extends the SongImporter and injects the Collab specific properties to SONG.

Additionally, the application.yml file and context objects from the core module were updated to rename all the SONG properties to be Collab properties.